Refactor models in common module
