+++ modules/common/src/main/scala/higherkindness/compendium/models/config/HttpConfig.scala
@@ -14,20 +14,6 @@
  * limitations under the License.
  */
 
-package higherkindness.compendium.http
+package higherkindness.compendium.models.config
 
-import higherkindness.compendium.models.{ErrorResponse, Protocol}
-import io.circe.{Decoder, Encoder}
-import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
-
-object Decoders {
-
-  implicit val protocolDecoder: Decoder[Protocol]           = deriveDecoder[Protocol]
-  implicit val errorResponseDecoder: Decoder[ErrorResponse] = deriveDecoder[ErrorResponse]
-}
-
-object Encoders {
-
-  implicit val protocolEnconder: Encoder[Protocol]          = deriveEncoder[Protocol]
-  implicit val errorResponseEncoder: Encoder[ErrorResponse] = deriveEncoder[ErrorResponse]
-}
+final case class HttpConfig(host: String, port: Int)
